# Optimize values() mongodb call