@CamronStaley 2025-04-17
adding my results with compression + lz4 `ds.values('filepath')`:

With these changes
```
(TEAMS) (3.10.14) ➜  scripts python benchmark.py -d malaria-cell-images -n 1000000 -r 3 -a retrieve_filepaths
Repetition 1
------------------------- Retrieve values (filepaths) --------------------------
Running...
Done in 17.412230968475342 seconds.

ACTION                       DATASET                CONNECTION_TYPE      VERSION  SAMPLE COUNT  TOTAL ELAPSED  ELAPSED PER SAMPLE
=================================================================================================================================
Retrieve values (filepaths)  malaria-cell-images            api_uri  2.7.0.dev29       1000000        17.4122              0.0000
Repetition 2
------------------------- Retrieve values (filepaths) --------------------------
Running...
Done in 17.00377106666565 seconds.

ACTION                       DATASET                CONNECTION_TYPE      VERSION  SAMPLE COUNT  TOTAL ELAPSED  ELAPSED PER SAMPLE
=================================================================================================================================
Retrieve values (filepaths)  malaria-cell-images            api_uri  2.7.0.dev29       1000000        17.0038              0.0000
Repetition 3
------------------------- Retrieve values (filepaths) --------------------------
Running...
Done in 16.805859088897705 seconds.

ACTION                       DATASET                CONNECTION_TYPE      VERSION  SAMPLE COUNT  TOTAL ELAPSED  ELAPSED PER SAMPLE
=================================================================================================================================
Retrieve values (filepaths)  malaria-cell-images            api_uri  2.7.0.dev29       1000000        16.8059              0.0000
```

without these changes:
```
(TEAMS) (3.10.14) ➜  scripts python benchmark.py -d malaria-cell-images -n 1000000 -r 3 -a retrieve_filepaths
Repetition 1
------------------------- Retrieve values (filepaths) --------------------------
Running...
Done in 21.160054922103882 seconds.

ACTION                       DATASET                CONNECTION_TYPE      VERSION  SAMPLE COUNT  TOTAL ELAPSED  ELAPSED PER SAMPLE
=================================================================================================================================
Retrieve values (filepaths)  malaria-cell-images            api_uri  2.7.0.dev29       1000000        21.1601              0.0000
Repetition 2
------------------------- Retrieve values (filepaths) --------------------------
Running...
Done in 21.426618814468384 seconds.

ACTION                       DATASET                CONNECTION_TYPE      VERSION  SAMPLE COUNT  TOTAL ELAPSED  ELAPSED PER SAMPLE
=================================================================================================================================
Retrieve values (filepaths)  malaria-cell-images            api_uri  2.7.0.dev29       1000000        21.4266              0.0000
Repetition 3
------------------------- Retrieve values (filepaths) --------------------------
Running...
Done in 21.77501916885376 seconds.

ACTION                       DATASET                CONNECTION_TYPE      VERSION  SAMPLE COUNT  TOTAL ELAPSED  ELAPSED PER SAMPLE
=================================================================================================================================
Retrieve values (filepaths)  malaria-cell-images            api_uri  2.7.0.dev29       1000000        21.7750              0.0000
(TEAMS) (3.10.14) ➜  scripts
```