@swheaton 2025-04-17
Ok I'll try it with non indexed scalar fields