@swheaton 2025-04-18
@kaixi-wang you're right, this is super weird. It's a particular behavior of MongoDB, with no clear explanation from the explain plan.

But what I've found boils down to the cardinality of the field.
Low cardinality or complex/embedded fields?: `_id: [False|True]` can make it MUCH slower
High cardinality: `_id: False` is 10-20% faster.

I'm thinking that we use this only for filepath and id/_id. That's really the only cases we can guarantee it will at least be no worse than before. Which will tie in with @brimoor's #5479